Updated App.js to fetch all pollutant data so it can be used across all views. Updated MainView and DetailedView to reflect this so the pollutants will appear in dropdown
