+++ dashboard/src/App.js
@@ -1,4 +1,4 @@
-import React from "react";
+import React, { useState, useEffect } from "react";
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 
 import "./App.css";
@@ -6,14 +6,26 @@ import "leaflet/dist/leaflet.css";
 import MainView from "./components/MainView/MainView";
 import DetailedView from "./components/DetailedView/DetailedView";
 import HistoricalView from "./components/HistoricalView/HistoricalView";
+import { getAllPollutants } from "./api";
 
 function App() {
+  const [allPollutants, setAllPollutants] = useState();
+
+  const fetchAllPollutants = () => {
+    getAllPollutants()
+      .then((response) => setAllPollutants(response.results));
+  }
+
+  useEffect(() => {
+    fetchAllPollutants();
+    console.log(allPollutants);
+  }, [])
   return (
     <Router>
       <Routes>
-        <Route path="/detailed" element={<DetailedView />} />
+        <Route path="/detailed" element={<DetailedView allPollutants={allPollutants} />} />
         <Route path="/historical" element={<HistoricalView />} />
-        <Route path="/" element={<MainView />} />
+        <Route path="/" element={<MainView allPollutants={allPollutants} />} />
       </Routes>
     </Router>
   );
